feat(cloud): raise a connect-compute notice on the notebook stage
A fresh cloud notebook has nowhere to run but only said so inside the
collapsed workstations rail. Raise that state onto the stage and move the
pairing surface into one shared WorkstationConnectDialog mounted above the
rail, so it survives a rail panel switch or collapse and the workstations
management page can reuse it.
