+++ apps/notebook-cloud/viewer/use-cloud-workstations.ts
@@ -178,7 +178,9 @@ export function useCloudWorkstationManager({
       onSetDefaultWorkstation: canChooseHostedWorkstation ? handleSetDefaultWorkstation : undefined,
       onStartPairing: canChooseHostedWorkstation ? handleStartPairing : undefined,
       onCancelPairing: canChooseHostedWorkstation ? handleCancelPairing : undefined,
+      canRegisterWorkstation: canChooseHostedWorkstation,
       workstationAction,
+      workstationLaunchReadiness,
       workstationPairing: canChooseHostedWorkstation ? pairingWithName : null,
       workstationPanelStatusMessage,
       workstationSelection,
@@ -194,6 +196,7 @@ export function useCloudWorkstationManager({
       pairingWithName,
       workstationSurface.busyWorkstationId,
       workstationAction,
+      workstationLaunchReadiness,
       workstationPanelStatusMessage,
       workstationSelection,
     ],
